+++ report-service/src/main/java/com/example/reportservice/dto/visit_auth/VisitAuthRequestResponseDTO.java
@@ -1,4 +1,4 @@
-package com.example.reportservice.dto;
+package com.example.reportservice.dto.visit_auth;
 
 
 import lombok.AllArgsConstructor;
